Completed settings:
- added dropdown for selecting settings tab on mobile
- corrected ui and design descrepancies
- refactored components to be <= 100 lines of code
