+++ app/(auth)/register/store-setup/page.tsx
@@ -12,9 +12,9 @@ import { useForm } from "react-hook-form";
 import { storeSetupSchemas } from "@/lib/validations/schemas";
 import { yupResolver } from "@hookform/resolvers/yup";
 import { storeSetupDefaultValues } from "@/lib/validations/defaults";
-import ConfirmAccountModal from "../../ui/register/storeSetup/ConfirmAccountModal";
+import ConfirmAccountModal from "../../ui/register/storeSetup/paymentOption/ConfirmAccountModal";
 import TermsOfContract from "../../ui/register/storeSetup/TermOfContract";
-import Stepper from "../../ui/register/storeSetup/Stepper";
+import Stepper from "../../ui/register/storeSetup/stepper/Stepper";
 import { useRouter } from "next/navigation";
 import { useStoreSetupContext } from "@/app/(auth)/contexts/storeSetupContext";
 import { ArrowBackButton } from "../../ui/buttons";
