+++ app/(auth)/ui/register/storeSetup/stepper/stepper-icons.tsx
