+++ app/dashboard/lib/interface.ts
@@ -1,0 +1,11 @@
+// ============================================================================
+// Navigation & UI Component Interfaces
+// ============================================================================
+
+export interface NavLink {
+    name: string;
+    href: string;
+    icon: string;
+    activeIcon: string;
+    active?: boolean;
+}
